Refactor chat UI: remove ChatBubble, extract InputBar, unify with TimelineEntry
- Delete ChatBubble component and extract InputBar + formatTime into dedicated InputBar.tsx
- Replace ChatBubble with TimelineEntry in project chat view for consistent rendering
- Add interaction dismiss (expire) functionality to global chat
- Widen project chat layout from max-w-3xl to max-w-4xl
- Add @floating-ui/react dependency
