+++ frontend/taskguild/src/components/organisms/TimelineEntry.tsx
@@ -7,7 +7,7 @@ import {
   Shield, MessageSquare, Bell, Mail, Play, Square, RefreshCw, Anchor,
   Terminal, AlertTriangle, Cog, Wrench, FileText, Zap, ChevronRight, ChevronDown,
 } from 'lucide-react'
-import { formatTime } from './ChatBubble.tsx'
+import { formatTime } from './InputBar.tsx'
 import { Badge } from '../atoms/index.ts'
 
 export type TimelineItem =
