+++ frontend/taskguild/src/routes/projects/$projectId/tasks/$taskId.tsx
@@ -15,7 +15,7 @@ import { useNotificationSound } from '@/hooks/useNotificationSound'
 import { TaskDetailModal } from '@/components/organisms/TaskDetailModal'
 import { ForceTransitionDialog } from '@/components/organisms/ForceTransitionDialog'
 import { shortId } from '@/lib/id'
-import { InputBar } from '@/components/organisms/ChatBubble'
+import { InputBar } from '@/components/organisms/InputBar'
 import { MarkdownDescription } from '@/components/organisms/MarkdownDescription'
 import { TimelineEntry, type TimelineItem } from '@/components/organisms/TimelineEntry'
 import { PendingRequestsPanel } from '@/components/organisms/PendingRequestsPanel'
